+++ src/AppserverIo/Http/HttpResponse.php
@@ -678,4 +678,19 @@ class HttpResponse implements ResponseInterface
         // Status-Line = HTTP-Version SP Status-Code SP Reason-Phrase CRLF
         return $this->getVersion() . ' ' . $this->getStatusCode() . ' ' . $this->getStatusReasonPhrase() . "\r\n";
     }
+
+    /**
+     * Redirects to the passed URL by adding a 'Location' header and
+     * setting the apropriate status code, by default 301.
+     *
+     * @param string  $url  The URL to forward to
+     * @param integer $code The status code to set
+     *
+     * @return void
+     */
+    public function redirect($url, $code = 301)
+    {
+        $this->setStatusCode($code);
+        $this->addHeader(HttpProtocol::HEADER_LOCATION, $url);
+    }
 }